Refactor the API code into its own app.
