+++ api/v1_0/serializers/publisher.py
