+++ api/views.py
@@ -6,25 +6,7 @@ from rest_framework.parsers import FormParser, MultiPartParser
 from rest_framework.permissions import IsAdminUser, IsAuthenticated
 from rest_framework.response import Response
 
-from comicsdb.filters.issue import IssueFilter
-from comicsdb.filters.name import ComicVineFilter, NameFilter, UniverseFilter
-from comicsdb.filters.series import SeriesFilter
-from comicsdb.models import (
-    Arc,
-    Character,
-    Creator,
-    Credits,
-    Imprint,
-    Issue,
-    Publisher,
-    Role,
-    Series,
-    Team,
-    Universe,
-)
-from comicsdb.models.series import SeriesType
-from comicsdb.models.variant import Variant
-from comicsdb.serializers import (
+from api.v1_0.serializers import (
     ArcListSerializer,
     ArcSerializer,
     CharacterListSerializer,
@@ -54,6 +36,24 @@ from comicsdb.serializers import (
     UniverseSerializer,
     VariantSerializer,
 )
+from comicsdb.filters.issue import IssueFilter
+from comicsdb.filters.name import ComicVineFilter, NameFilter, UniverseFilter
+from comicsdb.filters.series import SeriesFilter
+from comicsdb.models import (
+    Arc,
+    Character,
+    Creator,
+    Credits,
+    Imprint,
+    Issue,
+    Publisher,
+    Role,
+    Series,
+    Team,
+    Universe,
+)
+from comicsdb.models.series import SeriesType
+from comicsdb.models.variant import Variant
 from metron.throttle import GetUserRateThrottle, PostUserRateThrottle
